+++ curves/secp256k1/src/curves/tests.rs
@@ -2,3 +2,4 @@ use crate::Projective;
 use ark_algebra_test_templates::*;
 
 test_group!(g1; Projective; sw);
+test_group!(gl_glv; Projective; glv);