+++ curves/secq256k1/src/curves/mod.rs
@@ -1,8 +1,9 @@
+use ark_ec::scalar_mul::glv::GLVConfig;
 use ark_ec::{
     models::CurveConfig,
     short_weierstrass::{self as sw, SWCurveConfig},
 };
-use ark_ff::{AdditiveGroup, Field, MontFp, Zero};
+use ark_ff::{AdditiveGroup, BigInt, Field, MontFp, PrimeField, Zero};
 
 use crate::{fq::Fq, fr::Fr};
 
@@ -45,6 +46,32 @@ impl SWCurveConfig for Config {
     #[inline(always)]
     fn mul_by_a(_: Self::BaseField) -> Self::BaseField {
         Self::BaseField::zero()
+    }
+}
+
+impl GLVConfig for Config {
+    const ENDO_COEFFS: &'static [Self::BaseField] = &[MontFp!(
+        "78074008874160198520644763525212887401909906723592317393988542598630163514318"
+    )];
+    const LAMBDA: Self::ScalarField =
+        (MontFp!("60197513588986302554485582024885075108884032450952339817679072026166228089408"));
+    const SCALAR_DECOMP_COEFFS: [(bool, <Self::ScalarField as PrimeField>::BigInt); 4] = [
+        (false, BigInt!("64502973549206556628585045361533709078")),
+        (false, BigInt!("367917413016453100223835821029139468249")),
+        (false, BigInt!("303414439467246543595250775667605759171")),
+        (true, BigInt!("64502973549206556628585045361533709078")),
+    ];
+
+    fn endomorphism(p: &Projective) -> Projective {
+        let mut res = (*p).clone();
+        res.x *= Self::ENDO_COEFFS[0];
+        res
+    }
+
+    fn endomorphism_affine(p: &Affine) -> Affine {
+        let mut res = (*p).clone();
+        res.x *= Self::ENDO_COEFFS[0];
+        res
     }
 }